services/backend/pyproject.toml
[project]
name = "novareel-backend"
version = "0.1.0"
description = "FastAPI API and worker for NovaReel"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.35.0",
  "fastapi>=0.116.0",
  "httpx>=0.27.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.8.0",
  "PyJWT>=2.10.1",
  "cryptography>=44.0.0",
  "python-multipart>=0.0.20",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "ruff>=0.11.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app", "app.*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
